requirements.txt
python-multipart==0.0.6
Pillow==9.5.0
keras==2.12.0
tensorflow==2.12.0
uvicorn==0.22.0
fastapi==0.96.0